@@ -24,10 +24,10 @@
     }
   ],
   "require": {
-    "php": "^7.4 || ^8.0",
-    "bk2k/bootstrap-package": "^13.0",
-    "friendsoftypo3/headless": "^3.0.3",
-    "typo3/cms-core": "^11.5"
+    "php": "^8.1",
+    "bk2k/bootstrap-package": "^13.0 || ^14.0",
+    "friendsoftypo3/headless": "^4.2",
+    "typo3/cms-core": "^12.4"
   },
   "require-dev": {
     "ext-json": "*",
@@ -41,7 +41,7 @@
     "phpunit/phpcov": "^8.2",
     "saschaegerer/phpstan-typo3": "^1.8.2",
     "typo3/coding-standards": "^0.5.0",
-    "typo3/testing-framework": "^6"
+    "typo3/testing-framework": "^7"
   },
   "suggest": {
     "friendsoftypo3headless/headless-backendlayouts": "^1.0",
