@@ -53,7 +53,7 @@ class AccordionElementTest extends BaseContentTypeTest
         foreach ($contentElement['content']['items'] as $item) {
             self::assertArrayHasKey('id', $item);
             self::assertEquals('Header', $item['header'], 'accordion item: header mismatch');
-            self::assertEquals('<p><a href="/page1?parameter=999&amp;cHash=bfd4c1935d34c545ca918205373b0a42" title="LinkTitle" target="_blank" class="LinkClass">Link</a></p>', $item['bodytext']);
+            self::assertEquals('<p><a href="/page1?parameter=999&amp;cHash=bfd4c1935d34c545ca918205373b0a42" target="_blank" title="LinkTitle" class="LinkClass">Link</a></p>', $item['bodytext']);
             self::assertEquals('left', $item['mediaorient'], 'accordion item: mediaorient mismatch');
             self::assertEquals(2, $item['imagecols'], 'accordion item: imagecols mismatch');
             self::assertEquals(0, $item['imageZoom'], 'accordion item: imageZoom mismatch');
