@@ -61,7 +61,7 @@ class IconGroupElementTest extends BaseContentTypeTest
         self::assertCount(1, $contentElement['content']['items']);
 
         foreach ($contentElement['content']['items'] as $item) {
-            self::assertEquals('<p><a href="/page1?parameter=999&amp;cHash=bfd4c1935d34c545ca918205373b0a42" title="LinkTitle" target="_blank" class="LinkClass">Link</a></p>', $item['bodytext']);
+            self::assertEquals('<p><a href="/page1?parameter=999&amp;cHash=bfd4c1935d34c545ca918205373b0a42" target="_blank" title="LinkTitle" class="LinkClass">Link</a></p>', $item['bodytext']);
             self::assertEquals('Header', $item['header']);
             self::assertEquals('Subheader', $item['subheader']);
 
