@@ -29,7 +29,7 @@ class PanelElementTest extends BaseContentTypeTest
 
         // content element specific tests
         self::assertEquals('secondary', $contentElement['content']['panelClass'], 'panelClass mismatch');
-        self::assertEquals('<p><a href="/page1?parameter=999&amp;cHash=bfd4c1935d34c545ca918205373b0a42" title="LinkTitle" target="_blank" class="LinkClass">Link</a></p>', $contentElement['content']['bodytext'], 'bodytext mismatch');
+        self::assertEquals('<p><a href="/page1?parameter=999&amp;cHash=bfd4c1935d34c545ca918205373b0a42" target="_blank" title="LinkTitle" class="LinkClass">Link</a></p>', $contentElement['content']['bodytext'], 'bodytext mismatch');
         $this->checkDisabledFields($contentElement);
 
         // general tests
