@@ -28,7 +28,7 @@ class TextIconElementTest extends BaseContentTypeTest
         $contentElement = $fullTree['content']['colPos0'][18];
 
         // content element specific tests
-        self::assertEquals('<p><a href="/page1?parameter=999&amp;cHash=bfd4c1935d34c545ca918205373b0a42" title="LinkTitle" target="_blank" class="LinkClass">Link</a></p>', $contentElement['content']['bodytext'], 'bodytext mismatch');
+        self::assertEquals('<p><a href="/page1?parameter=999&amp;cHash=bfd4c1935d34c545ca918205373b0a42" target="_blank" title="LinkTitle" class="LinkClass">Link</a></p>', $contentElement['content']['bodytext'], 'bodytext mismatch');
         self::assertArrayNotHasKey('flexform', $contentElement);
         $this->checkIcon($contentElement);
 
